+++ src/main/java/com/cuupa/classificator/controller/GuiController.kt
@@ -16,10 +16,7 @@ import org.springframework.web.servlet.ModelAndView
 import java.time.LocalDate
 
 @Controller
-class GuiController(private val classificator: Classificator, private val manager: KnowledgeManager,
-                    private val monitor: Monitor) {
-
-    private val gson = Gson()
+class GuiController(private val classificator: Classificator, private val manager: KnowledgeManager) {
 
     @RequestMapping(value = ["/", "/index"], method = [RequestMethod.GET])
     fun index(model: Model): String {
@@ -40,38 +37,5 @@ class GuiController(private val classificator: Classificator, private val manage
         //manager.reloadKB()
         model.addAttribute("guiProcess", guiProcess)
         return "index"
-    }
-
-    @RequestMapping(value = ["/monitor"], method = [RequestMethod.GET])
-    fun monitor(model: Model): ModelAndView {
-        val monitorProcess = MonitorProcess()
-        monitorProcess.events = load(monitorProcess)
-        model.addAttribute("monitorProcess", monitorProcess)
-        model.addAttribute("data", gson.toJson(monitorProcess.events))
-        val modelAndView = ModelAndView("monitor")
-        val statistics = monitor.getStatistics(null, null)
-        modelAndView.addObject("topics", gson.toJson(statistics.topicDistribution))
-        modelAndView.addObject("senders", gson.toJson(statistics.senderDistribution))
-        return modelAndView
-    }
-
-    @RequestMapping(value = ["/monitorWithFilter"], method = [RequestMethod.POST])
-    fun monitorWithFilter(@ModelAttribute monitorProcess: MonitorProcess, model: Model): String {
-        monitorProcess.events = load(monitorProcess)
-        model.addAttribute("monitorProcess", monitorProcess)
-        return "monitor"
-    }
-
-    @RequestMapping(value = ["/exportAsPdf"], method = [RequestMethod.POST])
-    fun exportAsPdf(@ModelAttribute monitorProcess: MonitorProcess) {
-        val start: LocalDate? = monitorProcess.from
-        val end: LocalDate? = monitorProcess.to
-        val statistics = monitor.getStatistics(start, end)
-    }
-
-    private fun load(monitorProcess: MonitorProcess): List<Event> {
-        val start: LocalDate? = monitorProcess.from
-        val end: LocalDate? = monitorProcess.to
-        return monitor.getEvents(start, end)
     }
 }
